backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sse-starlette>=1.8.0
python-dotenv>=1.0.0
pydantic>=2.0.0
# 기존 dependencies (이미 설치됨)
langgraph>=0.6.0
litellm>=1.0.0
gspread>=5.0.0
google-auth>=2.0.0
pandas>=1.3.0
